pyproject.toml
[tool.poetry]
name = "vizwizard"
version = "0.1.0"
description = "Python dependencies for Viz Wizard"
authors = ["Casper Damen"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = "^1.4.1"
plotly = "^5.6.0"
openpyxl = "^3.0.9"
altair = "^4.2.0"
streamlit = "^1.8.1"
azure-identity = "^1.9.0"
azure-keyvault-secrets = "^4.4.0"
pyodbc = "^4.0.32"
jinjasql = "^0.1.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
